refactor: extract useNavigationHandlers from Dashboard
Extract touch swipe handlers and keyboard navigation useEffect into
useNavigationHandlers hook. Moves touch refs (touchStartX/Y, touchEndX/Y)
into the hook since they're only used by swipe handlers. Removes unused
useCallback import from Dashboard.

Dashboard.js: 2,010 → 1,614 lines (-396)
